+++ lib/core/color/app_color_manger.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class AppColorManger {
-  static const Color primary = Color(0xffFED377);
+  static const Color primary = Color(0xffF3AFA6);
   static const Color amber = Color(0xFFFFC107);
   static const Color gray = Color(0xFFA5A5A5);
 }
