+++ lib/core/theming/app_color.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
 
 class ColorManger {
-  static Color primary = const Color(0xF35CC9F0);
+  static Color primary = const Color(0xffF3AFA6);
 }
