+++ lib/features/onboarding/widgt/custom_elevated_button.dart
@@ -3,17 +3,19 @@ import 'package:flutter/material.dart';
 class CustomElevatedButton extends StatelessWidget {
   final String text;
   final Color color;
+  final void Function()? onPressed;
   const CustomElevatedButton({
     super.key,
     required this.text,
     required this.color,
+    this.onPressed,
   });
 
   @override
   Widget build(BuildContext context) {
     return ElevatedButton(
       style: ButtonStyle(backgroundColor: MaterialStatePropertyAll(color)),
-      onPressed: () {},
+      onPressed: onPressed,
       child: Text(text),
     );
   }
