+++ lib/main_app.dart
@@ -1,10 +1,11 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:installment/core/app_local.dart';
 import 'package:installment/core/routing/app_routers.dart';
 import 'package:installment/core/routing/routers.dart';
 import 'package:installment/generated/l10n.dart';
+
+import 'core/app_scroll_behavior.dart';
 
 class MainApp extends StatelessWidget {
   final AppRouter appRouter;
@@ -55,13 +56,4 @@ class MainApp extends StatelessWidget {
       ),
     );
   }
-}
-
-class AppScrollBehavior extends MaterialScrollBehavior {
-  @override
-  Set<PointerDeviceKind> get dragDevices => {
-        PointerDeviceKind.touch,
-        PointerDeviceKind.mouse,
-        PointerDeviceKind.trackpad,
-      };
 }
